Slim demo mode to a single home-page affordance
