+++ internal/demo/guard.go
@@ -1,7 +1,6 @@
 package demo
 
 import (
-	"html/template"
 	"log/slog"
 	"net/http"
 	"strings"
@@ -21,8 +20,8 @@ type Deps struct {
 }
 
 // Guard wraps the app handler with demo-mode behavior. When demo mode is off it
-// returns next unchanged (zero overhead). When on, it serves GET /demo and
-// POST /demo/enter, 404s the blocked paths, and delegates the rest.
+// returns next unchanged (zero overhead). When on, it serves POST /demo/enter,
+// 404s /profile and its subpaths, and delegates the rest.
 func Guard(next http.Handler, deps Deps) http.Handler {
 	if !Enabled() {
 		return next
@@ -33,7 +32,6 @@ func Guard(next http.Handler, deps Deps) http.Handler {
 		sessions: session.New([]byte(deps.Cfg.SessionKey), deps.Cfg.SecureCookies()),
 		players:  deps.Players,
 		logger:   deps.Logger,
-		tmpl:     template.Must(template.New("enter.gohtml").ParseFS(entryTmplFS, "templates/enter.gohtml")),
 	}
 }
 
@@ -42,15 +40,9 @@ type handler struct {
 	sessions *session.Manager
 	players  auth.PlayerStore
 	logger   *slog.Logger
-	tmpl     *template.Template
 }
 
 func (h *handler) ServeHTTP(w http.ResponseWriter, r *http.Request) {
-	if r.URL.Path == "/demo" && r.Method == http.MethodGet {
-		h.serveEntry(w, r)
-
-		return
-	}
 	if r.URL.Path == "/demo/enter" && r.Method == http.MethodPost {
 		h.enter(w, r)
 
@@ -67,15 +59,31 @@ func (h *handler) ServeHTTP(w http.ResponseWriter, r *http.Request) {
 
 // isBlocked reports whether path matches a blocked prefix exactly or as a
 // segment boundary (so /profile and /profile/password match, /profiles does not).
-// Blocked prefixes: account self-service (/profile/*) and real sign-in/signup
-// (/register, Google OAuth). Outbound email is not blocked here - the demo
-// deployment runs with SMTP unconfigured, so the app already uses a no-op mailer.
+// Blocked prefix: account self-service (/profile/*).
+// Outbound email is not blocked here - the demo deployment runs with SMTP
+// unconfigured, so the app already uses a no-op mailer.
 func isBlocked(path string) bool {
-	for _, p := range []string{"/profile", "/register", "/login/google"} {
+	for _, p := range []string{"/profile"} {
 		if path == p || strings.HasPrefix(path, p+"/") {
 			return true
 		}
 	}
 
 	return false
+}
+
+// enter logs the visitor into the shared demo Host by establishing a session
+// server-side (no password - the same mechanism OAuth uses), then redirects to
+// the host dashboard. CSRF is intentionally not required: it logs into a shared,
+// daily-wiped public account with no sensitive target.
+func (h *handler) enter(w http.ResponseWriter, r *http.Request) {
+	host, err := h.players.GetPlayerByDisplayName(r.Context(), demoHostName)
+	if err != nil {
+		h.logger.ErrorContext(r.Context(), "demo enter: host not ready", slog.Any("err", err))
+		http.Error(w, "demo host not ready", http.StatusServiceUnavailable)
+
+		return
+	}
+	h.sessions.Set(w, host.ID, host.SessionVersion)
+	http.Redirect(w, r, "/admin/quizzes", http.StatusSeeOther)
 }
