+++ internal/demo/guard_test.go
@@ -1,12 +1,15 @@
 package demo_test
 
 import (
+	"log/slog"
 	"net/http"
 	"net/http/httptest"
 	"testing"
 
 	"github.com/starquake/topbanana/internal/config"
+	"github.com/starquake/topbanana/internal/dbtest"
 	"github.com/starquake/topbanana/internal/demo"
+	"github.com/starquake/topbanana/internal/store"
 )
 
 func okHandler() http.Handler {
@@ -38,7 +41,7 @@ func TestGuard_BlocksLockedPaths(t *testing.T) {
 	t.Setenv("DEMO_MODE_ENABLED", "true")
 	g := demo.Guard(okHandler(), demo.Deps{Cfg: guardCfg()})
 
-	for _, path := range []string{"/profile", "/profile/password", "/register", "/login/google", "/login/google/callback"} {
+	for _, path := range []string{"/profile", "/profile/password"} {
 		req := httptest.NewRequestWithContext(t.Context(), http.MethodGet, path, nil)
 		rec := httptest.NewRecorder()
 		g.ServeHTTP(rec, req)
@@ -52,12 +55,44 @@ func TestGuard_PassesUnblocked(t *testing.T) {
 	t.Setenv("DEMO_MODE_ENABLED", "true")
 	g := demo.Guard(okHandler(), demo.Deps{Cfg: guardCfg()})
 
-	for _, path := range []string{"/", "/client/", "/login", "/admin/quizzes"} {
+	for _, path := range []string{"/", "/client/", "/login", "/admin/quizzes", "/register", "/login/google"} {
 		req := httptest.NewRequestWithContext(t.Context(), http.MethodGet, path, nil)
 		rec := httptest.NewRecorder()
 		g.ServeHTTP(rec, req)
 		if got, want := rec.Code, http.StatusOK; got != want {
 			t.Errorf("unblocked %s code = %d, want %d", path, got, want)
 		}
+	}
+}
+
+func TestGuard_EnterLogsInDemoHost(t *testing.T) {
+	t.Setenv("DEMO_MODE_ENABLED", "true")
+
+	logger := slog.New(slog.DiscardHandler)
+	stores := store.New(dbtest.Open(t), logger)
+	// Seed the host (reuses SeedIfEnabled's host path via a full seed).
+	cfg := demoTestConfig()
+	cfg.SessionKey = "test-session-key-test-session-ke"
+	if _, err := stores.Players.CreateAnonymousPlayer(t.Context(), "Demo Host"); err != nil {
+		t.Fatalf("seed host err = %v", err)
+	}
+
+	g := demo.Guard(okHandler(), demo.Deps{Cfg: cfg, Players: stores.Players, Logger: logger})
+
+	req := httptest.NewRequestWithContext(t.Context(), http.MethodPost, "/demo/enter", nil)
+	rec := httptest.NewRecorder()
+	g.ServeHTTP(rec, req)
+
+	if got, want := rec.Code, http.StatusSeeOther; got != want {
+		t.Fatalf("POST /demo/enter code = %d, want %d", got, want)
+	}
+	var sawSession bool
+	for _, c := range rec.Result().Cookies() {
+		if c.Name == "topbanana_session" && c.Value != "" {
+			sawSession = true
+		}
+	}
+	if !sawSession {
+		t.Error("POST /demo/enter set no topbanana_session cookie, want one")
 	}
 }
