+++ internal/web/tmpl/admin/layouts/base.gohtml
@@ -42,7 +42,6 @@
     <script>if('serviceWorker' in navigator)navigator.serviceWorker.register('/sw.js');</script>
 </head>
 <body class="bg-bg text-text font-sans antialiased min-h-dvh flex flex-col leading-relaxed">
-{{/* DEMO MODE */}}{{template "demo_banner" .}}
 {{template "topbar" .}}
 <section class="max-w-shell w-full mx-auto px-5 py-10 grow">
     {{template "content" .}}
