+++ internal/web/tmpl/auth/layouts/base.gohtml
@@ -44,7 +44,6 @@
     <script>if('serviceWorker' in navigator)navigator.serviceWorker.register('/sw.js');</script>
 </head>
 <body class="bg-bg text-text font-sans antialiased min-h-dvh leading-relaxed flex flex-col">
-{{/* DEMO MODE */}}{{template "demo_banner" .}}
 {{/* The auth card stays vertically centered in the content region. The
      client surface (#893) drops the shared topbar - the brand wordmark
      and account cross-links live in client_footer instead. */}}
