+++ internal/web/tmpl/home/layouts/base.gohtml
@@ -55,7 +55,6 @@
      short content (#285). dvh tracks the visible viewport across
      URL-bar collapse, eliminating the constant overflow. */}}
 <body class="bg-bg text-text font-sans antialiased min-h-dvh leading-relaxed flex flex-col">
-{{/* DEMO MODE */}}{{template "demo_banner" .}}
 {{/* The client surface (#893) doesn't carry the shared topbar - account
      controls live in client_footer. Distinguishes the player site from
      the admin console visually + structurally. */}}
