+++ internal/web/tmpl/home/pages/index.gohtml
@@ -1,4 +1,13 @@
 {{define "content"}}
+{{if demoMode}}
+<div class="mb-8 rounded-lg border border-accent-line bg-surface px-6 py-4 flex flex-col sm:flex-row sm:items-center justify-between gap-4">
+    <p class="m-0 text-sm text-text-dim">This is the Top Banana! demo - everything resets daily.</p>
+    <form method="POST" action="/demo/enter" class="m-0 shrink-0">
+        <input type="hidden" name="csrf_token" value="{{csrfToken}}">
+        <button type="submit" class="btn-primary">Build &amp; host a quiz</button>
+    </form>
+</div>
+{{end}}
     {{/* The brand logo lives in the shared top bar now (#844); the hero
          keeps only the page title + tagline so the wordmark is not
          duplicated. The <h1> stays the page's primary heading. */}}
