+++ internal/web/tmpl/host/layouts/base.gohtml
@@ -24,8 +24,6 @@
     <script defer src="/static/js/vendor/alpine.min.js"></script>
 </head>
 <body class="bg-bg text-text font-sans antialiased min-h-dvh">
-    {{/* DEMO MODE: deliberately NO demo_banner here - this is the projected
-         big-screen layout; the banner stays on the interactive picker (page.gohtml). */}}
     {{template "content" .}}
 </body>
 </html>
