+++ internal/web/tmpl/host/layouts/page.gohtml
@@ -11,7 +11,6 @@
     <link rel="stylesheet" href="/static/css/app.css">
 </head>
 <body class="bg-bg text-text font-sans antialiased min-h-dvh">
-    {{/* DEMO MODE */}}{{template "demo_banner" .}}
     {{/* Minimal host header: the wordmark links back to the dashboard, and a
          persistent "Session live" indicator (#889) appears when the host already
          has a room open so they know they are mid-session while picking, and can
